Rework around a 'state transition table' and rename files.
Rework TransitionAllowed() to use a state transition table to check valid transitions, to eliminate changing code each time a different state transition is needed ... now just a table adjustment is needed.
The code is done in a manner to execute efficiently through indexing and not searching for states, due to the index being used with the String() function.
Tests updated to cover all transitions.
Files renamed to explicitly identify which state is being adjusted as later on there will be a content collection with its own state - so constants also renamed to reflect this.