Get subset of query working with mongo
The original method did not work when just one or a few types of query was asked for - it resulted in no data returned from mongo as it looks like one can not pick and also exclude required types in the one built up `select`.
Also improve error response for unknown query string and update swagger.
Update tests.