Allow for more time to pass in service test between calling shutdown and
the service to be able to close all connections. This will enforce a forced
shutdown due to graceful shutdown timeout exceeding deadline needed for test.

Previously the shutdown would be graceful and other times not, this was solely to
do with the timeouts being too close to one another resulting in different outcomes
when running the unit test.